fix(android): newEngine() (minuscule Kotlin) + dstPort.toLong() (int Go -> long Java)
gomobile génère les constructeurs comme fonctions package-level en
minuscules côté Java/Kotlin: NewEngine() Go -> newEngine() Kotlin.
De plus int Go -> long Java: dstPort (Int) -> toLong() pour decide().

+++ android/app/src/main/java/com/dlnraja/faillefox/vpn/TunnelHandler.kt
@@ -61,8 +61,8 @@ class TunnelHandler(
             }
             val protoStr = when (proto) { 6 -> "tcp"; 17 -> "udp"; else -> "ip" }
 
-            // Décision du moteur Go.
-            val decision = engine.decide("android", protoStr, dstIp, dstPort)
+            // Décision du moteur Go. gomobile mappe int Go -> long Java/Kotlin.
+            val decision = engine.decide("android", protoStr, dstIp, dstPort.toLong())
             when (decision) {
                 "allow" -> { /* forward : v0.14 implémentera le NAT complet */ }
                 "deny"  -> { /* drop */ }
